requirements.txt
django
django-rest-framework
drf-spectacular
mux-python
pillow
python-dotenv
python-jose[cryptography]
requests